Split variable.cpp for parallel compilation.
Building this object takes extremely long due to many instantiations
related to transform.
